added ImGui Support + wip debugging GUI for PvZ

+++ src/Lawn/Challenge.cpp
@@ -2870,7 +2870,7 @@ void Challenge::WhackAZombieSpawning()
 					mBoard->GetTopPlantAt(aGridItem->mGridX, aGridItem->mGridY, TOPPLANT_ONLY_NORMAL_POSITION);
 				if (aPlant == nullptr || aPlant->mSeedType != SEED_GRAVEBUSTER)
 				{
-					aGridPicks[aGridPicksCount].mItem = (int)aGridItem;
+					aGridPicks[aGridPicksCount].mItem = (uintptr_t)aGridItem;
 					aGridPicks[aGridPicksCount].mWeight = 1;
 					aGridPicksCount++;
 				}
@@ -3295,7 +3295,7 @@ void Challenge::MoveAPortal()
 		{
 			TOD_ASSERT(aNumpicks < MAX_PORTALS);
 			aPickArray[aNumpicks].mWeight = 1;
-			aPickArray[aNumpicks].mItem = (int)aGridItem;
+			aPickArray[aNumpicks].mItem = (uintptr_t)aGridItem;
 			aNumpicks++;
 		}
 	}
@@ -3885,7 +3885,7 @@ void Challenge::ScaryPotterChangePotType(GridItemState thePotType, int theCount)
 			if ((thePotType == GRIDITEM_STATE_SCARY_POT_LEAF && aGridItem->mScaryPotType == SCARYPOT_SEED) ||
 				(thePotType == GRIDITEM_STATE_SCARY_POT_ZOMBIE && aGridItem->mZombieType == ZOMBIE_GARGANTUAR))
 			{
-				aPotArray[aPotCount].mItem = (int)aGridItem;
+				aPotArray[aPotCount].mItem = (uintptr_t)aGridItem;
 				aPotArray[aPotCount].mWeight = 1;
 				aPotCount++;
 			}
